Additional Test cases

+++ src/main/java/seedu/address/model/ModelManager.java
@@ -192,7 +192,7 @@ public class ModelManager implements Model {
      */
     public ObservableList<PieChart.Data> getIncomePieChartData() {
 
-        int[] incomePieChartArr = new int[6];
+        double[] incomePieChartArr = new double[6];
 
         for (Entry e : filteredIncome) {
             String tagName = e.getTag().getTagName();
@@ -225,7 +225,6 @@ public class ModelManager implements Model {
             new PieChart.Data("Profit", incomePieChartArr[2]), new PieChart.Data("Investment", incomePieChartArr[3]),
             new PieChart.Data("Gifts", incomePieChartArr[4]), new PieChart.Data("Others", incomePieChartArr[5]));
 
-
         return FXCollections.unmodifiableObservableList(incomePieChartData);
     }
 
@@ -236,7 +235,7 @@ public class ModelManager implements Model {
      * @return ObservableList of expense pie chart data
      */
     public ObservableList<PieChart.Data> getExpensePieChartData() {
-        int[] expensePieChartArr = new int[6];
+        double[] expensePieChartArr = new double[6];
 
         for (Entry e : filteredExpenditure) {
             String tagName = e.getTag().getTagName();
